requirements.txt
django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.4
celery==5.2.7
redis==4.3.4
django-environ==0.9.0
wkhtmltopdf==0.2